ocpp_16/requirements.txt
aiohttp ==3.8.4
anyio==3.6.2
asyncpg==0.27.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
databases==0.7.0
fastapi==0.89.1
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4

ocpp~=0.17.0